feat(reports):added export pdf option
and others functions like generate weekly, monthly or cutomized range dated reports in pdf, with a professional looking.

+++ app/schemas/report_schema.py
@@ -6,7 +6,7 @@ from pydantic import BaseModel
 
 class ReportCategory(BaseModel):
     category: str
-    net_category_balance: int
+    net_category_balance: float
 
 class ReportTransaction(BaseModel):
     id: int
